Add display joined mission in seperate component
