+++ src/pages/profile/Profile.jsx
@@ -1,35 +1,11 @@
 import React from 'react';
-import { useSelector } from 'react-redux';
-import Navbar from '../../components/navbar/Navbar';
-import { selectAllMissions } from '../../redux/missions/missionsSlice';
+import ProfileCard from './profileCard';
 import './profile.css';
 
 const Profile = () => {
-  const missions = useSelector(selectAllMissions);
-  const joinedMissions = missions.filter((mission) => mission.reserved === true);
-  if (joinedMissions.length === 0) {
-    return (
-      <div>
-        <Navbar />
-        <div className="row">
-          <p className="text-muted">No joined missions</p>
-        </div>
-      </div>
-    );
-  }
   return (
     <div>
-      <Navbar />
-      <div className="profile-section">
-        <h2>My Missions</h2>
-        <div className="name-mission">
-          {joinedMissions.map((mission) => (
-            <div className="row" key={mission.id}>
-              <p>{mission.missionName}</p>
-            </div>
-          ))}
-        </div>
-      </div>
+      <ProfileCard />
     </div>
   );
 };
